requirements.txt
fastapi==0.112.0
uvicorn==0.30.6
python-jobspy==1.1.80
pydantic==2.8.2
pandas==2.3.0